extract tests from `rl-commons` to `rl-tests`
