+++ crates/rl-tests/Cargo.toml
@@ -12,3 +12,4 @@ rl-parser.workspace = true
 rl-interpreter.workspace = true
 rl-utils.workspace = true
 rl-vm.workspace = true
+rl-commons.workspace = true
