+++ crates/rl-tests/tests/tests.rs
@@ -1,4 +1,5 @@
 mod common;
+mod commons;
 mod interpreter;
 mod lexer;
 mod parser;
